+++ Geoprofs/tests/Feature/wessam/AccounttoevoegenControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\wessam;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
