+++ Geoprofs/tests/Unit/damien/LoginControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\damien;
 
 use App\Models\User;
 use App\Models\UserInfo;
