+++ Geoprofs/tests/Unit/wessam/RegisterControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\wessam;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
